+++ _internal/escapeRegExp.js
