+++ _internal/isThenable.js
