+++ core/pipeR.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-const ensureExecutable = require('../internal/ensureExecutable')
+const ensureExecutable = require('../_internal/ensureExecutable')
 
 const pipeR = func => value => func (pipeR (func)).reduce (
   (acc, x) => acc.then (ensureExecutable (x)),
