+++ function/curry.js
@@ -1,9 +1,9 @@
 /* eslint-disable */
-const curry2 = require('../internal/curry2')
-const curry3 = require('../internal/curry3')
-const curry4 = require('../internal/curry4')
-const curry5 = require('../internal/curry5')
-const curry6 = require('../internal/curry6')
+const curry2 = require('../_internal/curry2')
+const curry3 = require('../_internal/curry3')
+const curry4 = require('../_internal/curry4')
+const curry5 = require('../_internal/curry5')
+const curry6 = require('../_internal/curry6')
 
 const curry = n =>
   n === 2 ? curry2
