+++ list/filter.js
@@ -1,6 +1,6 @@
 /* eslint-disable */
-const isThenable = require('../internal/isThenable')
-const iterableSerialReduce = require('../internal/iterableSerialReduce')
+const isThenable = require('../_internal/isThenable')
+const iterableSerialReduce = require('../_internal/iterableSerialReduce')
 
 const asyncFilterReducer = func => (acc, x) => {
   const result = func(x)
