+++ string/replace.js
@@ -1,5 +1,5 @@
 /* eslint-disable */
-const escapeRegExp = require('../internal/escapeRegExp')
+const escapeRegExp = require('../_internal/escapeRegExp')
 
 const replace = pattern => replacement => string =>
   string.replace (new RegExp (escapeRegExp (pattern), 'g'), replacement)
